pyproject.toml
[tool.poetry]
name = "webling-telegram-bot"
version = "0.1.0"
description = "Ein Telegram Bot, der offene Anträge aus Webling abfragt und in einen Telegram-Kanal postet"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.12"
requests = "*"
python-telegram-bot = "*"
python-dotenv = "*"
schedule = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
black = "*"
flake8 = "*"
mypy = "*"
isort = "*"
types-requests = "^2.32.4.20250611"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py312']

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 